MusChart (PC 0.4)

+++ muschart/src/main/java/by/gsu/database/dao/IRelationDAO.java
@@ -2,23 +2,23 @@ package by.gsu.database.dao;
 
 import java.util.List;
 
-import by.gsu.bean.Artist;
-import by.gsu.bean.Genre;
-import by.gsu.bean.Track;
 import by.gsu.exception.ValidationException;
+import by.gsu.model.Artist;
+import by.gsu.model.Genre;
+import by.gsu.model.Track;
 
 public interface IRelationDAO extends IDAO {
 
-    void updateTrackArtists(int idTrack, List<Artist> artists) throws ValidationException;
+    void updateTrackArtists(long idTrack, List<Artist> artists) throws ValidationException;
 
-    void updateTrackGenres(int idTrack, List<Genre> genres) throws ValidationException;
+    void updateTrackGenres(long idTrack, List<Genre> genres) throws ValidationException;
 
-    void updateArtistGenres(int idArtist, List<Genre> genres) throws ValidationException;
+    void updateArtistGenres(long idArtist, List<Genre> genres) throws ValidationException;
 
-    void updateUserTracks(int idUser, List<Track> tracks) throws ValidationException;
+    void updateUserTracks(long idUser, List<Track> tracks) throws ValidationException;
 
-    void updateUserArtists(int idUser, List<Artist> artists) throws ValidationException;
+    void updateUserArtists(long idUser, List<Artist> artists) throws ValidationException;
 
-    void updateUserGenres(int idUser, List<Genre> genres) throws ValidationException;
+    void updateUserGenres(long idUser, List<Genre> genres) throws ValidationException;
 
 }
